CRUD ops: create - prefix issue with issue type
