+++ dot-config/nvim/plugin/21_gh.lua
@@ -233,56 +233,8 @@ vim.api.nvim_create_user_command(
 vim.api.nvim_create_user_command(
   'IssueCreate',
   function()
-    -- Step 1: Select issue type
-    vim.ui.select(
-      { 'fix', 'enhancement' },
-      { prompt = 'Select issue type:' },
-      function(issue_type)
-        if not issue_type then return end
-        -- Step 2: Get issue title
-        vim.ui.input(
-          { prompt = 'Issue title: ' },
-          function(title)
-            if not title or title == '' then return end
-            -- Step 3: Collect checkbox list items
-            local tasks = {}
-            local function prompt_task()
-              vim.ui.input(
-                { prompt = 'Add task (empty to finish): ' },
-                function(task)
-                  if not task or task == '' then
-                    -- Assemble and create issue
-                    if #tasks == 0 then
-                      vim.notify('No tasks added, aborting issue creation', vim.log.levels.WARN)
-                      return
-                    end
-                    local body = table.concat(vim.tbl_map(function(t)
-                      return '- [ ] ' .. t
-                    end, tasks), '\n')
-                    local show = require('show')
-                    -- Create git branch with hyphenated title
-                    local branch_name = title:lower():gsub('%s+', '-'):gsub('[^%w%-]', '')
-                    -- Create GitHub issue
-                    local args = { 'gh', 'issue', 'create', '--title', '"' .. title .. '"', '--body', '"' .. body .. '"' }
-                    if issue_type then
-                      table.insert(args, '--label')
-                      table.insert(args, '"' .. issue_type .. '"')
-                    end
-                    show.shell('IssueCreate', table.concat(args, ' '))
-                  else
-                    -- Add task and prompt for another
-                    table.insert(tasks, task)
-                    prompt_task()
-                  end
-                end
-              )
-            end
-
-            prompt_task()
-          end
-        )
-      end
-    )
+    local issue = require('issues')
+    issue.create()
   end,
   { desc = 'Create GitHub issue with checkbox list' }
 )
